requirements.txt
djangorestframework_simplejwt==4.7.1
Django==3.1.7
django_model_utils==4.1.1
djangorestframework==3.12.4
